requirements.txt
pytest
selenium
Appium-Python-Client
allure-pytest>=2.5.1
